Move binary to dedicated folder
Moves the binary to a dedicated folder to distance it from the library
code. Alongside the binary, the `logger` module is moved out of
`src/utils` since its inclusion inside the library makes only little
sense. This closes #26.
