+++ src/bin/logger/comblog.rs
