+++ src/bin/logger/simplelog.rs
